backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["nemmip <julkazbs2000@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-restful = "^0.3.10"
sqlalchemy = "^2.0.35"
yoyo-migrations = "^9.0.0"
uuid = "^1.30"
psycopg2-binary = "^2.9.9"
marshmallow = "^3.22.0"
flask-restx = "^1.3.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

flask-cors = "^5.0.0"
uvicorn = "^0.32.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = 'src.main:run'
